Split mock apis
