+++ package.json
@@ -54,6 +54,7 @@
     "nodemon": "^2.0.20",
     "parcel": "^2.8.2",
     "prettier": "^2.8.1",
+    "process": "^0.11.10",
     "typescript": "^4.9.4"
   },
   "engines": {
